examples/mongodb_example/requirements.txt
python-pzem
motor
dnspython
python-decouple